+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## NEXT
+## 0.21.0 [2023-11-07]
 
 - Show a snack bar when the user request a new email to verify his email.
 - Adds the reset password page.